Restructuring
Large restructuring of the file structure.  New convention is to put all assembly files relating to operation up to the switch into 32 bit mode belong in boot.  Kernel C files belong in kernel.  When drivers or other components not belonging in either category come along new folders will be added.

Along with the restructure I also rewrote the makefile to reflect it.  Compilation appears to be the same result.  Further, a new file, screen.c, has been added which handles all screen IO.  It was included in this commit in order to test the ability of my new makefile to accomodate arbitrary C files in the kernel directory.