Testing preparation
In order to properly run test cases I have created two new files, test.c and test.h.  This is so that kernel.c won't have to contain testing code, thus volatility is only expected in test.c and h.  Because of the volatility they are not included in source control.